requirements.txt
fastapi
uvicorn
transformers
Pillow
python-multipart